refactor: move decrypt_session_key
