+++ src/composed/message/decrypt.rs
@@ -4,22 +4,7 @@ use zeroize::{Zeroize, ZeroizeOnDrop};
 use crate::crypto::sym::SymmetricKeyAlgorithm;
 use crate::errors::Result;
 use crate::packet::SymKeyEncryptedSessionKey;
-use crate::types::{EskType, Password, PkeskBytes, SkeskVersion};
-use crate::SignedSecretKey;
-
-/// Decrypts session key using secret key.
-pub fn decrypt_session_key(
-    locked_key: &SignedSecretKey,
-    key_pw: &Password,
-    values: &PkeskBytes,
-    typ: EskType,
-) -> Result<PlainSessionKey> {
-    debug!("decrypt session key");
-
-    locked_key.unlock(key_pw, |pub_params, priv_key| {
-        priv_key.decrypt(pub_params, values, typ, &*locked_key.public_key())
-    })
-}
+use crate::types::{Password, SkeskVersion};
 
 /// Decrypted session key.
 ///
